requirements.txt
django==1.6
djangorestframework
django-cities
psycopg2==2.5.1